Modified to AlbumCardController
